Always included memory storage in builds
Using the `postgres` feature removed the entire `mod memory` from the
build, same if you're using `--all-features` for testing. For now we
always include it and allow the unused module, this will properly run
clippy and check over the memory storage with `--all-features`.

In the future we want something more robust to handle this..