Add option to destination to forward query parameters
Skip query params that are already in the location, params from the
location are leading. Overwriting this could result is problematic
redirects. Adding query params might already be an issue in some cases,
the redirect location should be able to handle the extra params, this is
why the option to append them is behind an option per destination.